requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/requirements.in
#
add-trailing-comma==2.3.0
    # via -r requirements/requirements.in
aiogram==2.22.1
    # via -r requirements/requirements.in
aiohttp==3.8.3
    # via aiogram
aioredis==2.0.1
    # via -r requirements/requirements.in
aiosignal==1.2.0
    # via aiohttp
alembic==1.8.1
    # via -r requirements/requirements.in
async-timeout==4.0.2
    # via
    #   aiohttp
    #   aioredis
    #   redis
asyncpg==0.26.0
    # via -r requirements/requirements.in
attrs==22.1.0
    # via aiohttp
babel==2.9.1
    # via aiogram
build==0.8.0
    # via pip-tools
certifi==2022.9.24
    # via aiogram
charset-normalizer==2.1.1
    # via aiohttp
click==8.1.3
    # via pip-tools
deprecated==1.2.13
    # via redis
flake8==5.0.4
    # via
    #   -r requirements/requirements.in
    #   flake8-quotes
flake8-quotes==3.3.1
    # via -r requirements/requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==1.1.3
    # via -r requirements/requirements.in
idna==3.4
    # via yarl
isort==5.10.1
    # via -r requirements/requirements.in
loguru==0.6.0
    # via -r requirements/requirements.in
mako==1.2.3
    # via alembic
markupsafe==2.1.1
    # via mako
mccabe==0.7.0
    # via flake8
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via
    #   build
    #   redis
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r requirements/requirements.in
psycopg2==2.9.3
    # via -r requirements/requirements.in
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyparsing==3.0.9
    # via packaging
python-decouple==3.6
    # via -r requirements/requirements.in
pytz==2022.6
    # via babel
redis==4.3.4
    # via -r requirements/requirements.in
sqlalchemy==1.4.41
    # via
    #   -r requirements/requirements.in
    #   alembic
tokenize-rt==5.0.0
    # via add-trailing-comma
tomli==2.0.1
    # via
    #   build
    #   pep517
typing-extensions==4.4.0
    # via aioredis
wheel==0.38.2
    # via pip-tools
wrapt==1.14.1
    # via deprecated
yarl==1.8.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
